refactor: SearchBar -> Header로 rename
